services/auth-service/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-JWT-Extended
bcrypt
python-dotenv
redis
requests
flask-sqlalchemy
